requirements.txt
streamlit
pandas
scikit-learn
numpy
plotly
langchain
langchain-groq
python-dotenv
streamlit-mic-recorder
gtts
speechrecognition
pyaudio